power: oplus_vooc: Condtionally skip reading header/footer from vooc fw
On OOS, the kernel is handed a firmware payload with the 80 byte header and 128
byte footer removed; the header/footer is likely removed by some userspace
service. On custom ROMs, that service does not exist, so the kernel falls back
to reading the firmware files directly from `/odm/firmware/fastchg`, complete
with the header and footer. However, since the kernel still assumes that the
header and footer are removed, trying to write this firmware to the charging IC
fails.

Fix this by checking for the presence of `FwUp` in the beginning of the
firmware payload. If it's present, we can then assume that the header and
footer is still in the payload, and thus do some pointer arithmetic and
subtract from the size field to exclude the header and footer. This will allow
VOOC firmware uploading to work on both OOS and custom ROMs.

Co-authored-by: Abhay Singh Gill <abhaygill017@gmail.com>
Suggested-by: jabashque <jabashque@gmail.com>
Signed-off-by: Abhay Singh Gill <abhaygill017@gmail.com>